restructured some comps
