+++ src/app/(root)/watchlist/page.tsx
@@ -1,6 +1,9 @@
 import React from 'react'
+import { getWatchlistByUserId } from '@/lib/actions/watchlist.actions'
 
-const WatchlistPage = () => {
+const WatchlistPage = async () => {
+  const watchlist = await getWatchlistByUserId()
+
   return (
     <div>WatchlistPage</div>
   )
